pyproject.toml
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
allow_redefinition = true
show_error_codes = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "transformers.*",
    "PIL.*",
    "torch.*",
    "reverse_geocoder.*",
    "pytest.*",
    "piexif.*",
    "typer.*",
    "rich.*",
    "fastapi.*",
    "dotenv.*",
    "uvicorn.*"
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disable_error_code = [
    "no-untyped-def",
    "no-untyped-call",
    "operator",
    "index",
    "assignment",
    "arg-type",
    "no-any-return"
]
check_untyped_defs = false
disallow_untyped_defs = false

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-image-organizer"
version = "0.1.0"
description = "Smart Image Organizer - Automatically organize images using metadata and AI"
dependencies = [
    "typer",
    "rich",
    "fastapi",
    "python-dotenv",
    "uvicorn",
    "pillow",
    "requests",
    "python-jose",
    "passlib",
    "types-requests",
    "types-python-jose",
    "types-passlib",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=src --cov-report=html --cov-report=term-missing"

[coverage.run]
source = ["src"]
omit = ["tests/*", "setup.py"]

[coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "pass",
]
